requirements.txt
cycler==0.11.0
fonttools==4.28.1
imageio==2.11.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.2.0
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
